+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## X.Y.Z 2022-XX-XX
+## 22.7.1 2022-08-31
 ### PaymentSheet
 * [Fixed] PaymentSheet now uses configuration.apiClient for Apple Pay instead of always using STPAPIClient.shared
 * [Fixed] Fixed Mac Catalyst support in Xcode 14. [#2001](https://github.com/stripe/stripe-ios/issues/2001)